backend/requirements.txt
# Eventually Yours Shopping App - Backend Dependencies
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
httpx==0.24.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
